+++ packages/backup/service_test.go
@@ -6,10 +6,13 @@ import (
 
 	"github.com/cloudogu/ces-control-api/generated/backup"
 	backupV1 "github.com/cloudogu/k8s-backup-lib/api/v1"
+	componentV1 "github.com/cloudogu/k8s-component-lib/api/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
 func Test_getAllBackups(t *testing.T) {
@@ -198,6 +201,210 @@ func Test_getAllRestores(t *testing.T) {
 		// then
 		require.NoError(t, err)
 		assert.Equal(t, 0, len(allRestores.Restores))
+	})
+}
+
+func TestDefaultBackupService_GetSchedule(t *testing.T) {
+	testCtx := context.Background()
+	schedule := &backupV1.BackupSchedule{Spec: backupV1.BackupScheduleSpec{Schedule: "1 2 * 4 *"}}
+
+	t.Run("should get schedule", func(t *testing.T) {
+		mBackupScheduleClient := newMockBackupScheduleClient(t)
+		mBackupScheduleClient.EXPECT().Get(testCtx, "ces-schedule", metav1.GetOptions{}).Return(schedule, nil)
+
+		svc := &DefaultBackupService{
+			backupScheduleClient: mBackupScheduleClient,
+		}
+
+		response, err := svc.GetSchedule(testCtx, nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, "1 2 * 4 *", response.Schedule)
+	})
+
+	t.Run("should fail to get empty schedule with error", func(t *testing.T) {
+		mBackupScheduleClient := newMockBackupScheduleClient(t)
+		mBackupScheduleClient.EXPECT().Get(testCtx, "ces-schedule", metav1.GetOptions{}).Return(nil, assert.AnError)
+
+		svc := &DefaultBackupService{
+			backupScheduleClient: mBackupScheduleClient,
+		}
+
+		_, err := svc.GetSchedule(testCtx, nil)
+
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to get backup schedule:")
+	})
+}
+
+func TestDefaultBackupService_SetSchedule(t *testing.T) {
+	testCtx := context.Background()
+
+	t.Run("should set schedule", func(t *testing.T) {
+		expectedSchedule := &backupV1.BackupSchedule{
+			ObjectMeta: metav1.ObjectMeta{Name: "ces-schedule"},
+			Spec:       backupV1.BackupScheduleSpec{Schedule: "* 2 3 * *"},
+		}
+
+		mBackupScheduleClient := newMockBackupScheduleClient(t)
+		mBackupScheduleClient.EXPECT().Get(testCtx, "ces-schedule", metav1.GetOptions{}).Return(nil, k8sErrors.NewNotFound(schema.GroupResource{}, "not found"))
+		mBackupScheduleClient.EXPECT().Create(testCtx, expectedSchedule, metav1.CreateOptions{}).Return(expectedSchedule, nil)
+
+		svc := &DefaultBackupService{
+			backupScheduleClient: mBackupScheduleClient,
+		}
+
+		response, err := svc.SetSchedule(testCtx, &backup.SetBackupScheduleRequest{Schedule: "* 2 3 * *"})
+
+		require.NoError(t, err)
+		assert.Equal(t, &backup.SetBackupScheduleResponse{}, response)
+	})
+
+	t.Run("should fail to set schedule", func(t *testing.T) {
+		mBackupScheduleClient := newMockBackupScheduleClient(t)
+		mBackupScheduleClient.EXPECT().Get(testCtx, "ces-schedule", metav1.GetOptions{}).Return(nil, assert.AnError)
+
+		svc := &DefaultBackupService{
+			backupScheduleClient: mBackupScheduleClient,
+		}
+
+		_, err := svc.SetSchedule(testCtx, &backup.SetBackupScheduleRequest{Schedule: "* 2 3 * *"})
+
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to get existing backup schedule:")
+	})
+}
+
+func TestDefaultBackupService_GetRetentionPolicy(t *testing.T) {
+	testCtx := context.Background()
+
+	t.Run("should get default policy", func(t *testing.T) {
+		mComponentClient := newMockComponentClient(t)
+		mComponentClient.EXPECT().Get(testCtx, "k8s-backup-operator", metav1.GetOptions{}).Return(&componentV1.Component{
+			Spec: componentV1.ComponentSpec{},
+		}, nil)
+
+		svc := &DefaultBackupService{
+			componentClient: mComponentClient,
+		}
+
+		response, err := svc.GetRetentionPolicy(testCtx, nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, backup.RetentionPolicy_RETENTION_POLICY_UNSPECIFIED, response.Policy)
+	})
+
+	t.Run("should get policy", func(t *testing.T) {
+		mComponentClient := newMockComponentClient(t)
+		mComponentClient.EXPECT().Get(testCtx, "k8s-backup-operator", metav1.GetOptions{}).Return(&componentV1.Component{
+			Spec: componentV1.ComponentSpec{
+				ValuesYamlOverwrite: `
+cleanup:
+  exclude: foo
+retention:
+  strategy: "removeAllButKeepLatest"
+  garbageCollectionCron: "0 * * * *"
+`,
+			},
+		}, nil)
+
+		svc := &DefaultBackupService{
+			componentClient: mComponentClient,
+		}
+
+		response, err := svc.GetRetentionPolicy(testCtx, nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, backup.RetentionPolicy_RETENTION_POLICY_REMOVE_ALL_BUT_KEEP_LATEST, response.Policy)
+	})
+
+	t.Run("should get policy keepAll", func(t *testing.T) {
+		mComponentClient := newMockComponentClient(t)
+		mComponentClient.EXPECT().Get(testCtx, "k8s-backup-operator", metav1.GetOptions{}).Return(&componentV1.Component{
+			Spec: componentV1.ComponentSpec{
+				ValuesYamlOverwrite: `
+cleanup:
+  exclude: foo
+retention:
+  strategy: "keepAll"
+  garbageCollectionCron: "0 * * * *"
+`,
+			},
+		}, nil)
+
+		svc := &DefaultBackupService{
+			componentClient: mComponentClient,
+		}
+
+		response, err := svc.GetRetentionPolicy(testCtx, nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, backup.RetentionPolicy_RETENTION_POLICY_KEEP_ALL, response.Policy)
+	})
+
+	t.Run("should get policy keepLastSevenDays", func(t *testing.T) {
+		mComponentClient := newMockComponentClient(t)
+		mComponentClient.EXPECT().Get(testCtx, "k8s-backup-operator", metav1.GetOptions{}).Return(&componentV1.Component{
+			Spec: componentV1.ComponentSpec{
+				ValuesYamlOverwrite: `
+cleanup:
+  exclude: foo
+retention:
+  strategy: "keepLastSevenDays"
+  garbageCollectionCron: "0 * * * *"
+`,
+			},
+		}, nil)
+
+		svc := &DefaultBackupService{
+			componentClient: mComponentClient,
+		}
+
+		response, err := svc.GetRetentionPolicy(testCtx, nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, backup.RetentionPolicy_RETENTION_POLICY_KEEP_LAST_SEVEN_DAYS, response.Policy)
+	})
+
+	t.Run("should get policy keep7Days1Month1Quarter1Year", func(t *testing.T) {
+		mComponentClient := newMockComponentClient(t)
+		mComponentClient.EXPECT().Get(testCtx, "k8s-backup-operator", metav1.GetOptions{}).Return(&componentV1.Component{
+			Spec: componentV1.ComponentSpec{
+				ValuesYamlOverwrite: `
+cleanup:
+  exclude: foo
+retention:
+  strategy: "keep7Days1Month1Quarter1Year"
+  garbageCollectionCron: "0 * * * *"
+`,
+			},
+		}, nil)
+
+		svc := &DefaultBackupService{
+			componentClient: mComponentClient,
+		}
+
+		response, err := svc.GetRetentionPolicy(testCtx, nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, backup.RetentionPolicy_RETENTION_POLICY_KEEP_LAST_7_DAYS_OLDEST_OF_1_MONTH_1_QUARTER_1_HALF_YEAR_1_YEAR, response.Policy)
+	})
+
+	t.Run("should fail to get empty policy with error", func(t *testing.T) {
+		mComponentClient := newMockComponentClient(t)
+		mComponentClient.EXPECT().Get(testCtx, "k8s-backup-operator", metav1.GetOptions{}).Return(nil, assert.AnError)
+
+		svc := &DefaultBackupService{
+			componentClient: mComponentClient,
+		}
+
+		_, err := svc.GetRetentionPolicy(testCtx, nil)
+
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to get backup-operator component:")
 	})
 }