+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added 
 - [#68] Get backups and restores from cluster
+- [#76] get & set backup-schedule
+- [#76] get retention-policy
 - [#77] Create backups in ui
 
 ## [v1.7.2] - 2025-09-17